requirement.txt
numpy
torch
torchvision
opencv_python
Pillow
stable_baselines3
gym
airsim
